platform/lab-controller/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
bcrypt
slowapi
pyyaml
apscheduler
anthropic
openai
